add souvenir api test

+++ app/Http/Controllers/SouvenirController.php
@@ -52,7 +52,6 @@ class SouvenirController extends Controller
         $souvenir->users()->attach($user->id, [
             'role' => 'admin',
             'joined_at' => now(),
-            'can_edit' => true
         ]);
 
         return response()->json($souvenir, 201);
@@ -85,23 +84,48 @@ class SouvenirController extends Controller
             return response()->json(['message' => 'Souvenir introuvable'], 404);
         }
 
+        $role = $souvenir->users
+            ->firstWhere('id', $user->id)?->pivot->role;
+
+        if ($role !== "admin") {
+            return response()->json(['message' => 'Vous n\'avez pas les permissions nécessaire pour faire cela'], 403);
+        }
+
         $validated = $request->validate([
-            'name' => 'sometimes|string|max:255',
-            'cover_image' => 'sometimes|nullable|string',
-            'is_closed' => 'sometimes|boolean',
-            'users' => 'sometimes|array', // si tu modifies les rôles
+            'title' => 'sometimes|string|max:255',
+            'cover_image' => 'sometimes|nullable|file|mimes:png,jpg,jpeg|max:10240',
+            'users' => 'sometimes|array',
         ]);
+
+        if ($request->hasFile('cover_image')) {
+            $path = $request->file('cover_image')->store('souvenirs', 'public');
+            $validated['cover_image'] = $path;
+        }
 
         $souvenir->update($validated);
 
-        return response()->json($souvenir, 200);
+        return response()->json([
+            'id' => $souvenir->id,
+            'title' => $souvenir->title,
+            'cover_image' => $souvenir->cover_image,
+        ], 200);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function delete(Souvenir $souvenir): JsonResponse
+    public function delete(Request $request, Souvenir $souvenir): JsonResponse
     {
+        $user = $request->user();
+
+        $souvenir->load('users');
+
+        $role = $souvenir->users
+            ->firstWhere('id', $user->id)?->pivot->role;
+
+        if ($role !== "admin") {
+            return response()->json(['message' => 'Vous n\'avez pas les permissions nécessaire pour faire cela'], 403);
+        }
         $souvenir->delete();
 
         return response()->json(['message' => 'souvenir supprimé']);
